zavrseni autoservisi i usluge, prepravljen bekend da se ne otvara sa https nego sa http, dodali scroll na autoservis detalje, podesili izgled proizvoda, povukli cijenu sa popustom i popust, u bekendu treba na update proizvoda dodati i unosenje popusta, kada se proizvod update-a popust nije dobar

+++ CarCareHub_/UI_/flutter_mobile/lib/models/product.dart
@@ -8,7 +8,7 @@ class Product {
   String? naziv;
   String? sifra;
   double? cijena;
-  double? popust;
+  int? popust;
   String? slika;
   double? cijenaSaPopustom;
   String? originalniBroj;
